refactored to implement permission system
The permissions system includes the CHAP authentication protocol, the
group membership attribute on all users, the "gate" module, and the
"factory" module which are used together to control access to the
datastore on a per-user leve.  None of this is really enforced, it's
just an elaborate information hiding scheme to help prevent programming
mistakes that may expose more data than we intend to.
 On branch master
	new file:   gae_py/factory.py
	new file:   gae_py/gate.py
	new file:   gae_py/groups.yaml
	new file:   gae_py/request.py
	renamed:    gae_py/users.py -> gae_py/session.py
	new file:   gae_py/store.py
	modified:   gae_py/users.py
	modified:   test/testrunner.py
	modified:   test/tools.py